+++ src/viperx/constants.py
@@ -22,6 +22,7 @@ TESTS_DIR = "tests"
 # Templates
 TEMPLATE_DIR_NAME = "templates"
 TEMPLATES_DIR = Path(__file__).parent / TEMPLATE_DIR_NAME
+USER_TEMPLATES_DIR = Path.home() / ".config" / "viperx" / "templates"
 
 # Types
 TYPE_CLASSIC = "classic"